mcp/tutorial/pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.1",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
]
